feat: ✨ agregar endpoint histories/store-session y caso de uso StoreSessionCommandHandler
- Se añadió el endpoint histories/store-session.
- Se implementó el caso de uso StoreSessionCommandHandler.
- Se renombraron las entidades, configuraciones, DbSet y clases relacionadas para eliminar el prefijo 'Irrigation' de History.
- Asegurar consistencia en los nombres y mejorar la claridad en el dominio.

+++ src/TechNovaLab.Irrigo.Application/Features/CropManagement/CreateCropType/CropTypeResponse.cs
@@ -3,5 +3,5 @@
     public sealed record CropTypeResponse(
         Guid PublicId,
         string Name,
-        int WaterRequiredPerDay);
+        double WaterRequiredPerDay);
 }
